webui/requirements.txt
# FedTracker WebUI Dependencies

gradio>=4.0.0
matplotlib>=3.5.0
numpy>=1.21.0
torch>=1.13.0
Pillow>=9.0.0